A fairly heavy bit of modding. have implemented the player class mostly and set up the game w/ a base level of reinforcements for game state 3. current game states are as follows:
state 1: waiting for go button to be hit. this is where the opening menu and game options are going to be set.

state 2: Selecting intial territories to control (comp still grabs random)

state 3: spend initial reinforcements on owned territories

state 4: view existing territories. the main game logic will need to go here

need to work more on making the borders not look like crap. need to make a background water tile to draw behind everything else. next major step is to write a logic function for player class that will attempt to play a whole turn. this will have to be done in stages so that some of it (like attacks and the like) can be blitted to the screen so that the player knows something is happening. will work on that in the short future.